requirements.txt
pandas>=1.3
numpy>=1.19
scikit-learn>=1.0
matplotlib>=3.3
seaborn>=0.11
scipy>=1.7
statsmodels>=0.13
opencv-python>=4.5
pillow>=8.0
pytesseract>=0.3
ydata-profiling>=4.2  # pandas-profiling replacement
imbalanced-learn>=0.9
